# Conflicts:
#	app/src/main/java/com/videobridge/app/AppContainer.kt
#	app/src/main/java/com/videobridge/app/MainActivity.kt
#	app/src/main/java/com/videobridge/data/proxy/LocalProxyServer.kt
#	app/src/main/java/com/videobridge/permissions/AndroidNetworkPermissionManager.kt
#	app/src/main/java/com/videobridge/playback/PlaybackEngine.kt
#	app/src/main/java/com/videobridge/ui/MainViewModel.kt

+++ app/src/main/java/com/videobridge/core/metadata/MetadataSanitizer.kt
@@ -42,7 +42,7 @@ object MetadataSanitizer {
 
             val type = Character.getType(codePoint)
             if (
-                type == Character.CONTROL ||
+                type == Character.CONTROL.toInt() ||
                 Character.isWhitespace(codePoint) ||
                 Character.isSpaceChar(codePoint)
             ) {
